agente_ia/backend/requirements.txt
# === AGENTE IA - DEPENDÊNCIAS BACKEND ===
# Core Python Libraries
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Machine Learning (Fase 2 - IA Híbrida Adaptativa)
scikit-learn>=1.3.0
tensorflow>=2.13.0
joblib>=1.3.0

# Data Processing & Analysis
watchdog>=3.0.0
psutil>=5.9.0

# SSH Remote Collection
paramiko>=3.0.0

# Email Alerts

# API & Web Framework
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Configuration & Logging
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
colorama>=0.4.6
tqdm>=4.65.0
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0 